Filter worklogs by name

Filter the elements of a worklog tree by name. For filter_name_keep, all worklog tree elements of the appropriate type with names that are matched by any of the elements of pattern are kept, minus any worklog tree elements with names that are matched by any of the elements of exclude. Subtrees that don't contain any worklogs after filtering can optionally be pruned from the tree.

Usage

filter_name_keep(wkls, pattern, type, exclude, prune_empty, ...)

# S4 method for worklogs_node
filter_name_keep(
  wkls,
  pattern = character(0L),
  type = "leafs",
  exclude = character(0L),
  prune_empty = TRUE,
  ...
)

Arguments

wkls

A worklogs object.

pattern

A character vector of regular expressions that will be used to match against the worklogs_node and worklog_leaf names.

type

One of either "leafs", "nodes", or "both" specifying which types of elements in the worklogs tree to perform matching upon.

exclude

A character vector of regular expressions that will be used to match against the worklogs_node and worklog_leaf names.

prune_empty

Either TRUE or FALSE specifying whether to prune any subtrees that don't contain worklogs in them after filtering.

...

Arguments that are passed on to base::grepl to be used for regular expression matching with pattern and exclude.

Details

All worklog tree elements (i.e. nodes and leafs) have an associated name, with the exception of the top level. filter_name_keep provides the ability to filter elements of the worklog tree based on their names and types.

filter_name_keep performs filtering by walking the worklog tree. For a given subtree in the tree, children nodes and children leafs are treated differentially according to the input to the type argument, as described in the following section.

Filtering algorithm

If type is "nodes":

For a node: when a node name is matched then include the entire subtree rooted by the node; otherwise recurse into the node and perform filtering on the node children.

For a leaf: the leaf is discarded

If type is "both":

For a node: when a node name is matched then include the entire subtree rooted by the node; otherwise recurse into the node and perform filtering on the node children.

For a leaf: keep the leaf if the name is a match, otherwise the leaf is discarded.

If type is "leaf":

For a node: recurse into the node and perform filtering on the node children.

For a leaf: keep the leaf if the name is a match, otherwise the leaf is discarded.

Regular expression pattern matching

When performing filtering, for a given input and tree node element the names of the node children are matched against the regular expressions provided for the pattern and exclude arguments.

Regular expression pattern matching works the same for the input to pattern as for exclude as follows. For each worklog child name of the relevant type (as determined by the input to the type argument), if any of the regular expressions provided as elements of the argument input character vector match the worklog child name then that worklog child is marked as being included or excluded, as appropriate.

Regular expression pattern matching is performed by the base::grepl function. You can use ... to provide arguments that will be passed on to grepl.